Split fetcher out to separate module
This paves the way for a future async fetcher that uses aiohttp instead
of requests, at which point requests would be an optional dependency.
